SimpleJWTAuthentication/requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.85.0
greenlet==1.1.3
h11==0.13.0
idna==3.4
pydantic==1.10.2
PyJWT==2.5.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing-extensions==4.3.0
uuid==1.30
uvicorn==0.18.3
